requirements.txt
requests
logging
werkzeug
sqlalchemy
flask
wtforms
dnspython
alembic
filetype
pillow

# app plugins
flask_sqlalchemy
flask_migrate
flask_login
flask_mail
